Improve Measure optional return values

+++ src/core/csr/measures/mapping_quality_zero_count.cpp
@@ -14,7 +14,7 @@ std::unique_ptr<Measure> MappingQualityZeroCount::do_clone() const
 
 Measure::ResultType MappingQualityZeroCount::do_evaluate(const VcfRecord& call, const FacetMap& facets) const
 {
-    return std::stoi(call.info_value("MQ0").front());
+    return static_cast<std::size_t>(std::stoull(call.info_value("MQ0").front()));
 }
 
 std::string MappingQualityZeroCount::do_name() const
